+++ avalon-modules/avalon-dataset/avalon-dataset-service/src/main/kotlin/io/github/lishangbu/avalon/dataset/service/impl/EncounterConditionValueServiceImpl.kt
@@ -6,6 +6,7 @@ import io.github.lishangbu.avalon.dataset.entity.dto.EncounterConditionValueView
 import io.github.lishangbu.avalon.dataset.entity.dto.SaveEncounterConditionValueInput
 import io.github.lishangbu.avalon.dataset.entity.dto.UpdateEncounterConditionValueInput
 import io.github.lishangbu.avalon.dataset.repository.EncounterConditionValueRepository
+import io.github.lishangbu.avalon.dataset.repository.LocationAreaEncounterConditionValueRepository
 import io.github.lishangbu.avalon.dataset.service.EncounterConditionValueService
 import org.babyfish.jimmer.sql.ast.mutation.SaveMode
 import org.springframework.stereotype.Service
@@ -15,15 +16,19 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class EncounterConditionValueServiceImpl(
     private val encounterConditionValueRepository: EncounterConditionValueRepository,
+    private val locationAreaEncounterConditionValueRepository: LocationAreaEncounterConditionValueRepository,
 ) : EncounterConditionValueService {
     @Transactional(rollbackFor = [Exception::class])
     override fun save(command: SaveEncounterConditionValueInput): EncounterConditionValueView = encounterConditionValueRepository.save(command.toEntity(), SaveMode.INSERT_ONLY).let(::reloadView)
 
     @Transactional(rollbackFor = [Exception::class])
-    override fun update(command: UpdateEncounterConditionValueInput): EncounterConditionValueView = encounterConditionValueRepository.save(command.toEntity(), SaveMode.UPSERT).let(::reloadView)
+    override fun update(command: UpdateEncounterConditionValueInput): EncounterConditionValueView = encounterConditionValueRepository.save(command.toEntity(), SaveMode.UPDATE_ONLY).let(::reloadView)
 
     @Transactional(rollbackFor = [Exception::class])
     override fun removeById(id: Long) {
+        if (locationAreaEncounterConditionValueRepository.existsByEncounterConditionValueId(id)) {
+            throw IllegalStateException("当前遭遇条件值已被地点区域遭遇引用，请先删除相关遭遇引用后再删除当前遭遇条件值")
+        }
         encounterConditionValueRepository.deleteById(id)
     }
